+++ src/data/tipssystem.cpp
@@ -25,25 +25,25 @@ void UpdateTipRecords() {
          "{:s}: tips system not implemented, doing nothing\n", __func__);
 }
 
-void SetTipLockedState(int id, bool state) {
+void SetTipLockedState(size_t id, bool state) {
   if (Implementation) return Implementation->SetTipLockedState(id, state);
   ImpLog(LogLevel::Warning, LogChannel::VMStub,
          "{:s}: tips system not implemented, doing nothing\n", __func__);
 }
 
-void SetTipUnreadState(int id, bool state) {
+void SetTipUnreadState(size_t id, bool state) {
   if (Implementation) return Implementation->SetTipUnreadState(id, state);
   ImpLog(LogLevel::Warning, LogChannel::VMStub,
          "{:s}: tips system not implemented, doing nothing\n", __func__);
 }
 
-void SetTipNewState(int id, bool state) {
+void SetTipNewState(size_t id, bool state) {
   if (Implementation) return Implementation->SetTipNewState(id, state);
   ImpLog(LogLevel::Warning, LogChannel::VMStub,
          "{:s}: tips system not implemented, doing nothing\n", __func__);
 }
 
-bool GetTipLockedState(int id) {
+bool GetTipLockedState(size_t id) {
   if (Implementation) return Implementation->GetTipLockedState(id);
   ImpLog(LogLevel::Warning, LogChannel::VMStub,
          "{:s}: tips system not implemented, returning false\n", __func__);
@@ -57,14 +57,14 @@ std::vector<TipsDataRecord>* GetTipRecords() {
   return nullptr;
 }
 
-TipsDataRecord* GetTipRecord(int id) {
+TipsDataRecord* GetTipRecord(size_t id) {
   if (Implementation) return &Implementation->Records[id];
   ImpLog(LogLevel::Warning, LogChannel::VMStub,
          "{:s}: tips system not implemented, returning NULL\n", __func__);
   return nullptr;
 }
 
-int GetTipCount() {
+size_t GetTipCount() {
   if (Implementation) return Implementation->TipEntryCount;
   ImpLog(LogLevel::Warning, LogChannel::VMStub,
          "{:s}: tips system not implemented, returning 0\n", __func__);
